ai-service/requirements.txt
fastapi>=0.103.2
uvicorn>=0.23.2
pydantic>=2.4.2
pandas>=2.1.1
scikit-learn>=1.3.1
python-dotenv>=1.0.0
python-dotenv>=1.0.0

